Fixes `create_index.py` bug
If the index already existed, the `create_index.py` script would
delete the index before importing. However, if it did not already
exist before trying to drop it, an error occurred.

+++ create_index.py
@@ -29,7 +29,8 @@ if __name__ == '__main__':
 
     setup_connection()
 
-    Verse._index.delete()
+    if Verse._index.exists():
+        Verse._index.delete()
     Verse.init()
 
     for verse in verses:
